ajustando projeto e criando controllers adicionado valid dep

+++ src/main/java/br/com/clinicawszd/clinicawszd/repository/AgendamentoRepository.java
@@ -3,5 +3,5 @@ package br.com.clinicawszd.clinicawszd.repository;
 import br.com.clinicawszd.clinicawszd.model.Agendamento;
 import org.springframework.data.repository.CrudRepository;
 
-public interface AgendamentoRepository extends CrudRepository<Agendamento, Integer> {
+public interface AgendamentoRepository extends CrudRepository<Agendamento, Long> {
 }
